Move stuff from RestfulBase to their new homes.
