+++ plugins/restful/RestfulBase.php
@@ -493,55 +493,6 @@ abstract class RestfulBase extends \RestfulPluginBase implements \RestfulInterfa
     return $options;
   }
 
-
-  /**
-   * Call the output format on the given data.
-   *
-   * @param array $data
-   *   The array of data to format.
-   *
-   * @return string
-   *   The formatted output.
-   */
-  public function format(array $data) {
-    return $this->formatterManager->format($data, $this->getPluginKey('formatter'));
-  }
-
-  /**
-   * Execute a user callback.
-   *
-   * @param mixed $callback
-   *   There are 3 ways to define a callback:
-   *     - String with a function name. Ex: 'drupal_map_assoc'.
-   *     - An array containing an object and a method name of that object.
-   *       Ex: array($this, 'format').
-   *     - An array containing any of the methods before and an array of
-   *       parameters to pass to the callback.
-   *       Ex: array(array($this, 'processing'), array('param1', 2))
-   * @param array $params
-   *   Array of additional parameters to pass in.
-   *
-   * @return mixed
-   *   The return value of the callback.
-   *
-   * @throws \RestfulException
-   */
-  public static function executeCallback($callback, array $params = array()) {
-    if (!is_callable($callback)) {
-      if (is_array($callback) && count($callback) == 2 && is_array($callback[1])) {
-        // This code deals with the third scenario in the docblock. Get the
-        // callback and the parameters from the array, merge the parameters with
-        // the existing ones and call recursively to reuse the logic for the
-        // other cases.
-        return static::executeCallback($callback[0], array_merge($params, $callback[1]));
-      }
-      $callback_name = is_array($callback) ? $callback[1] : $callback;
-      throw new \RestfulException(format_string('Callback function: @callback does not exists.', array('@callback' => $callback_name)));
-    }
-
-    return call_user_func_array($callback, $params);
-  }
-
   /**
    * Return the resource name.
    *
@@ -822,20 +773,6 @@ abstract class RestfulBase extends \RestfulPluginBase implements \RestfulInterfa
     }
 
     return $selected_controller;
-  }
-
-  /**
-   * Helper method to know if the current request is for a list.
-   *
-   * @return boolean
-   *   TRUE if the request is for a list. FALSE otherwise.
-   */
-  public function isListRequest() {
-    if ($this->getMethod() != \RestfulInterface::GET) {
-      return FALSE;
-    }
-    $path = $this->getPath();
-    return empty($path) || strpos($path, ',') !== FALSE;
   }
 
   /**
